requirements.txt
pandas
numpy 
scikit-learn
tqdm
mlflow
PyYAML
shap
xgboost
lightgbm
matplotlib
seaborn
mlxtend
streamlit
plotly